+++ src/gConsent/constants.ts
@@ -37,6 +37,11 @@ export const GC_CONSENT_STATUS_EXPLICITLY_GIVEN =
 export const GC_CONSENT_STATUS_REQUESTED =
   "https://w3id.org/GConsent#ConsentStatusRequested";
 
+export const GC_CONSENT_STATUS_DENIED_ABBREV = "ConsentStatusDenied";
+export const GC_CONSENT_STATUS_EXPLICITLY_GIVEN_ABBREV =
+  "ConsentStatusExplicitlyGiven";
+export const GC_CONSENT_STATUS_REQUESTED_ABBREV = "ConsentStatusRequested";
+
 export const SOLID_VC_ISSUER = "http://www.w3.org/ns/solid/terms#vcIssuer";
 
 // TODO: Add dependency on generated vocabulary.
@@ -84,8 +89,19 @@ export const ACCESS_CREDENTIAL_TYPE = new Set([
   CREDENTIAL_TYPE_LEGACY_CONSENT_REQUEST,
 ]);
 
-export const CONSENT_STATUS = new Set([
-  GC_CONSENT_STATUS_DENIED,
-  GC_CONSENT_STATUS_EXPLICITLY_GIVEN,
-  GC_CONSENT_STATUS_REQUESTED,
-] as const);
+export const ACCESS_GRANT_STATUS = Object.freeze(
+  new Set([
+    GC_CONSENT_STATUS_DENIED,
+    GC_CONSENT_STATUS_EXPLICITLY_GIVEN,
+    GC_CONSENT_STATUS_DENIED_ABBREV,
+    GC_CONSENT_STATUS_EXPLICITLY_GIVEN_ABBREV,
+  ])
+);
+
+export const ACCESS_REQUEST_STATUS = Object.freeze(
+  new Set([GC_CONSENT_STATUS_REQUESTED, GC_CONSENT_STATUS_REQUESTED_ABBREV])
+);
+
+export const ACCESS_STATUS = Object.freeze(
+  new Set([...ACCESS_GRANT_STATUS, ...ACCESS_REQUEST_STATUS])
+);